pyproject.toml
[build-system]
requires = ["setuptools>=69"]
build-backend = "setuptools.build_meta"

[project]
name = "idor-lens"
version = "0.1.0"
description = "IDOR detector via role/token differential testing + minimal proof + regression mode"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
dependencies = [
  "requests==2.32.3",
  "PyYAML==6.0.2",
]

[project.scripts]
idor-lens = "idor_lens.cli:main"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
strict = true
warn_unused_configs = true
mypy_path = "src"
